I added feature for startup defaultPlaylist parameter.
on start up add playlist name to command line i.e. node app.js kanye
If there is a default playlist associated with the string kanye inside of setPlaylist.js then that will be the default
 playlist. Otherwise it will set the default as normal (hot 100).

If an invalid playlist name is selected than it goes to hot 100.
If no playlist is selected it goes to hot 100.